+++ tests/RunTest.ps1
@@ -1,0 +1,8 @@
+#invoke Pester v5 test
+#requires -module Pester
+
+$config = New-PesterConfiguration
+
+$config.Output.Verbosity ="detailed"
+
+Invoke-Pester -Configuration $config